refactor: separate footer and waves
