+++ apps/ui/src/components/Layout.tsx
@@ -10,6 +10,7 @@ import { MultiConnectButton } from "./ConnectButton";
 import { CustomHeaderLink } from "./CustomHeaderLink";
 import { CustomHeaderLogo } from "./CustomHeaderLogo";
 import { EnvSelector } from "./EnvSelector";
+import { Footer } from "./Footer";
 import { Waves } from "./Waves";
 
 import "./Layout.scss";
@@ -20,37 +21,40 @@ export const Layout = ({
   readonly children?: ReactNode;
 }): ReactElement => {
   return (
-    <Waves>
-      <EuiSpacer />
-      <EuiSpacer />
-      <EuiHeader
-        position="fixed"
-        sections={[
-          {
-            items: [
-              <CustomHeaderLogo key="custom-header-logo" />,
-              <EuiHeaderLinks key="eui-header-links">
-                <CustomHeaderLink to={"/swap"}>Swap</CustomHeaderLink>
-                <CustomHeaderLink to={"/pools"}>Pools</CustomHeaderLink>
-                {/* TODO: Enable when token is launched */}
-                {/* <CustomHeaderLink to={"/stake"}>Stake</CustomHeaderLink> */}
-                <CustomHeaderLink to={"/help"}>Help</CustomHeaderLink>
-              </EuiHeaderLinks>,
-            ],
-            borders: "right",
-          },
-          {
-            items: [
-              <EuiHeaderSectionItem key="eui-header-section-item">
-                <MultiConnectButton size="s" fullWidth />
-              </EuiHeaderSectionItem>,
-              <EnvSelector key="env-selector" />,
-            ],
-            borders: "none",
-          },
-        ]}
-      />
-      {children}
-    </Waves>
+    <>
+      <Waves>
+        <EuiSpacer />
+        <EuiSpacer />
+        <EuiHeader
+          position="fixed"
+          sections={[
+            {
+              items: [
+                <CustomHeaderLogo key="custom-header-logo" />,
+                <EuiHeaderLinks key="eui-header-links">
+                  <CustomHeaderLink to={"/swap"}>Swap</CustomHeaderLink>
+                  <CustomHeaderLink to={"/pools"}>Pools</CustomHeaderLink>
+                  {/* TODO: Enable when token is launched */}
+                  {/* <CustomHeaderLink to={"/stake"}>Stake</CustomHeaderLink> */}
+                  <CustomHeaderLink to={"/help"}>Help</CustomHeaderLink>
+                </EuiHeaderLinks>,
+              ],
+              borders: "right",
+            },
+            {
+              items: [
+                <EuiHeaderSectionItem key="eui-header-section-item">
+                  <MultiConnectButton size="s" fullWidth />
+                </EuiHeaderSectionItem>,
+                <EnvSelector key="env-selector" />,
+              ],
+              borders: "none",
+            },
+          ]}
+        />
+        {children}
+      </Waves>
+      <Footer />
+    </>
   );
 };
